Add limit parameter to get_transactions_by_address and Tron staking support
Introduces an optional limit parameter to the get_transactions_by_address method across all chain providers and implementations, enabling transaction queries with result limits. Removes deprecated Tron and HyperCore providers from gem_chain_rpc, and refactors Tron client and models to support staking, transaction mapping, and integration with TronGrid. Also includes test improvements and minor code cleanups.

+++ crates/gem_xrp/src/rpc/client.rs
@@ -133,7 +133,7 @@ impl<C: Client> XRPClient<C> {
         Ok(result.result.ledger)
     }
 
-    pub async fn get_account_transactions(&self, address: String, limit: u32) -> Result<AccountLedger, Box<dyn Error + Send + Sync>> {
+    pub async fn get_account_transactions(&self, address: String, limit: usize) -> Result<AccountLedger, Box<dyn Error + Send + Sync>> {
         let params = json!({
             "method": "account_tx",
             "params": [
